\name{AMOR-package}
\alias{AMOR-package}
\alias{AMOR}
\docType{package}

\title{Abundance Matrix Operations in R}

\description{Collection of functions useful for manipulating and visualizing abundance matrix liek the ones commonly found in ecology.}

\details{
\tabular{ll}{
Package: \tab AMOR\cr
Type: \tab Package\cr
Version: \tab 0.0-12\cr
Date: \tab 2015-02-07\cr
License: \tab GPL-3\cr
}
}

\author{Sur from Dangl Lab}

\keyword{ package }
