% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.r
\name{beta_diversity}
\alias{beta_diversity}
\title{Beta diversity}
\usage{
beta_diversity(X, method = "Bray-Curtis", sample_dim = 2)
}
\arguments{
\item{X}{Matrix with of abundances}

\item{method}{Distance method to be used. One of: "Bray-Curtis"}

\item{sample_dim}{Dimension that contains samples, 1 for rows
and 2 for columns. The other dimension should contain the taxa}
}
\value{
A \code{dist} object
}
\description{
Calculate beta diversity from abundance matrix
}
\examples{
mat <- matrix(1:10,ncol = 2)
beta_diversity(mat)
}
\seealso{
\link{dist}
}
\author{
Sur Herrera Paredes
}
