% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_glm.r
\name{bootstrap_glm}
\alias{bootstrap_glm}
\alias{bootstrap_glm.default}
\alias{bootstrap_glm.Dataset}
\title{Bootstrap estimation of Generalized Linear Models on a matrix of observations}
\usage{
bootstrap_glm(x, ...)

\method{bootstrap_glm}{default}(
  x,
  Map,
  formula,
  family = poisson(link = "log"),
  response.name = "Count",
  N = 100,
  verbose = FALSE,
  ...
)

\method{bootstrap_glm}{Dataset}(
  x,
  formula,
  family = poisson(link = "log"),
  response.name = "Count",
  N = 100,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Either an abundance matrix with samples as columns
or a Dataset object}

\item{...}{Extra parameters for \link{matrix_glm}, which will in turn
pass them to \link{glm}. Typically it is seful to specify the 'control'
option.}

\item{Map}{A mapping file with sample metadata if x is a matrix. A data.frame
containing the variables to be modelled as columns and samples as rows.
The rows should be named with sample IDs and must correspond to the column names
from x if an abundance matrix was passed.}

\item{formula}{The formula for the GLM model.Only the right hand side of the equation
must be passed, and anything on the left side wild be silently ignored.}

\item{family}{The model family. See help of \link{glm} and \link{family}
for details}

\item{response.name}{The name to give to the response variable}

\item{N}{The number of bootstrap pseudoreplicates}

\item{verbose}{Logical. If true each call to matrix_glm will print
progress information.}
}
\value{
A bootglm object which is a list containing the following elements:
\describe{
\item{orig}{A \code{matrix.glm} object containing the output of \code{matrix.glm()}
on the original data.}
\item{boot}{A list containing elements: 1) coefficients, a three-dimensioanl
array containing the matrix of coefficients for each bootstrap pseudoreplicate;
and 2) SE, a three-dimensional array containing the matrix of standard errors for
each bootstrap pseudoreplicate.}
\item{coefficients}{A matrix of coefficients calculated as the mean accross each
bootstrap pseudoreplicates.}
\item{call}{Function call via match.call().}
\item{family}{GLM model family. An object of class \code{family}.}
\item{pseudoreplicates}{Matrix of dimensions (number iof samples) x (number of
pseudoreplicates), where the i-column gives the sample ID, of the samples used for
the i-th pseudoreplicate.}
\item{N}{Number of bootstrap pseudoreplicates performed.}
}
}
\description{
Uses bootstrap to estimate variance of \link{matrix_glm}
estimates. Takes a matrix of observations, and fits the specified
GLM on each species. It perfomrs N bootstrap pesudoreplicates and
returns the matrix of coefficients, for the original dataset as well
as bootstrap estimates. It uses glm_matrix() to fit each of the pseudoreplicates.
}
\details{
This works by calling matrix_glm() on the original data, and each pf the N
bootstrap pseudoreplicates generated.
}
\author{
Sur Herrera Paredes
}
