% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_by_taxonomy.r
\name{collapse_by_taxonomy}
\alias{collapse_by_taxonomy}
\alias{collapse_by_taxonomy.default}
\alias{collapse_by_taxonomy.Dataset}
\title{Collapse by taxonomy}
\usage{
collapse_by_taxonomy(...)

\method{collapse_by_taxonomy}{default}(Tab, Tax, Group = NULL, level = 4,
  FUN = sum, sepchar = ";")

\method{collapse_by_taxonomy}{Dataset}(Dat, Group = NULL, level = 4,
  FUN = sum, sepchar = ";")
}
\arguments{
\item{Tab}{An abundance matrix with samples as columns and taxa as rows}

\item{Tax}{A Taxonomy data.frame as defined by \link{create_dataset}}

\item{Group}{If NULL, then the "Taxonomy" column of the Taxonomy data.frame will be used.
If not NULL, it can be either a column number or name (i.e. character string), indicating
which column of the Taxonomy data.frame to use as grouping factor. It can also be a factor
or vector that can be coerced into factor.}

\item{level}{If Group is NULL, the taxonomy level to use from the Taxonomy column.}

\item{FUN}{Which function to use to aggregate}

\item{sepchar}{Separator character for taxonomy levels.}
}
\value{
The default method returns a numeric matrix.

The Dataset method returns a Dataset object when Dat includes a Map
element (see \code{\link{create_dataset}}). When the Map element is
missing it returns a \code{matrix} object.
}
\description{
Collapses taxa according to the taxonomy object or a given factor
}
\examples{
library(AMOR)
data(Rhizo)
data(Rhizo.map)
data(Rhizo.tax)

Dat <- create_dataset(Tab = Rhizo, Map = Rhizo.map, Tax = Rhizo.tax)
Dat.phyl <- collapse_by_taxonomy(Dat = Dat, level = 4)

# The followwing returns a matrix
Dat.collapsed1 <- collapse_by_taxonomy(Dat$Tab, Dat$Tax)

# The following returns a dataset object
Dat.collapsed2 <- collapse_by_taxonomy(Dat=Dat)
}
\seealso{
\link{create_dataset}
}
\author{
Sur Herrera Paredes
}
