% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_good_samples.r
\name{find_good_samples}
\alias{find_good_samples}
\alias{findGoodSamples}
\alias{Find}
\alias{samples}
\alias{with}
\alias{minumum}
\alias{number}
\alias{of}
\alias{reads}
\title{Find good samples}
\usage{
find_good_samples(x, min_counts)
}
\arguments{
\item{x}{Abundance matrix with samples as
columns}

\item{min_counts}{Minimum count per sample to keep}
}
\value{
A boolean vector specifying which samples to keep
}
\description{
Find good samples
}
\examples{
mat <- matrix(rep(c(2,0,1),each = 10), ncol = 3)
mat[ , find_good_samples(mat, 1)]
}
\author{
Sur Herrera Paredes
}
