% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_good_samples.r
\name{findGoodSamples}
\alias{findGoodSamples}
\alias{find_good_samples}
\alias{Find}
\alias{samples}
\alias{with}
\alias{minumum}
\alias{number}
\alias{of}
\alias{reads}
\title{Find good samples}
\usage{
findGoodSamples(x, min_counts)
}
\arguments{
\item{x}{Abundance matrix with samples as
columns}

\item{min_counts}{Minimum count per sample to keep}
}
\value{
A boolean vector specifying which samples to keep
}
\description{
Find good samples
}
