% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tax_level.r
\name{get_tax_level}
\alias{get_tax_level}
\title{Get tax level}
\usage{
get_tax_level(Tax, level = 4, sepchar = ";")
}
\arguments{
\item{Tax}{A data.frame with a column named 'Taxonomy' which is
a character vector with taxonomy strings.}

\item{level}{The level to return}

\item{sepchar}{The separator of each level in the
taxonomy string}
}
\value{
A vector of taxonomy strings truncated at the
specified level
}
\description{
Processes a vector of txonomy strings and
returns a vector with the specified level
}
\details{
It splits each taxonomy string according to separator and
returns the specified level. If there are fewer levels
than the specified number, it returns the value 'unclassified'
}
\examples{
tax <- data.frame(ID = letters[1:4],
                  Taxonomy = c('a;b;c;d','a;b','A;B;C','A;B;C;D'))
get_tax_level(tax, level = 2)
}
\author{
Sur Herrera Paredes
}
