% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_glmNB.r
\name{matrix_glmNB}
\alias{matrix_glmNB}
\alias{matrix_glmNB.default}
\alias{matrix_glmNB.Dataset}
\title{Fit Negative Binomial Generalized Linear Models on a matrix
of observations}
\usage{
matrix_glmNB(x, ...)

\method{matrix_glmNB}{default}(
  x = NULL,
  Map = NULL,
  formula = NULL,
  response.name = "Count",
  verbose = FALSE,
  ...
)

\method{matrix_glmNB}{Dataset}(x, formula = NULL, response.name = "Count", verbose = FALSE, ...)
}
\arguments{
\item{x}{Either a matrix of abundances with samples as columns and
species as rows or a Dataset object}

\item{...}{Other parameters for the \link{glm.nb} function like control.}

\item{Map}{A data.frame containing the variables to be modelled as
columns and samples as rows. The rows should be named with sample
IDs and must correspond to the column names from x if an abundance
matrix was passed}

\item{formula}{A formula specifying the model to be fit. Only the
right hand side of the equation must be passed, and anything on the
left side wild be silently ignored}

\item{response.name}{String indicating the name to be used for the
response (dependent) variable in the GLM}

\item{verbose}{Logical value. If true the taza name is printed while
the fit is in progress.}
}
\value{
Returns a matrix.glm object which is a list containing the following elements:
\describe{
  \item{coefficients}{A matrix of coefficients for the glm fit. The matix has dimensions S x p, where S is the number of species in the abundance matrix, and p the number of parameters in the specified GLM. Each row corresponds to an independent fit of the model with glm() function and the specified formula}
  \item{SE}{Similar to Coef, but this matrix contains the estimated Standard Errors (ie. the Standard deviation of the coefficients).}
  \item{AIC}{A named vector containing the Akaike Information Criteria (AIC) of each independent fit. The names of the vector correspond to the species IDs in the input abundance matrix}
  \item{theta}{A vector of theta (ie. the overdispersion parameter) for each taxon}
  \item{SE.theta}{A vector of standard errorrs for the overdispersion parameter.}
  \item{call}{Function call via match.call}
  \item{family}{GLM model family. An object of class \code{family}}
  \item{X}{Design matrix of the model fit.}
}
}
\description{
Takes a matrix of observations, and fits the specified Negative
Binomial GLM on each species, returning the matrix of coefficients.
}
\details{
This function will take each species (row) in the abundance matrix
and fit the specified (via the formula option) Negative binomial GLM
independently on each one, using the glm.nb function from then MASS
package.
}
\examples{
data(Rhizo)
data(Rhizo.map)
data(Rhizo.tax)
Dat <- create_dataset(Rhizo,Rhizo.map,Rhizo.tax)

m1.nb <- matrix_glmNB(x=Dat$Tab,Map=Dat$Map,
                      formula=~accession + plate)
m1.nb <- matrix_glmNB(Dat,formula=~accession + plate,
                      control=glm.control(epsilon=1e-5,maxit=25),
                      verbose=TRUE)
}
\seealso{
\link{matrix_glm}
}
\author{
Sur Herrera Paredes
}
