% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmd.R
\name{mlmd}
\alias{mlmd}
\title{Maximum Likelihood Model Dependence Simulations}
\usage{
mlmd(data, dep, treat, predictors, level, n.sims, method = "odds",
  matched.data = NULL)
}
\arguments{
\item{data}{A binary dependent data set.}

\item{dep}{The dependent outcome binary variable that will be used to measure the effect (e.g. "success").}

\item{treat}{The treatment variable (e.g. "treatment").}

\item{predictors}{Vector of predictor variables.}

\item{level}{Level of allowed interactions. Level = (1, 2, 3) and correspond to one-way, two-way and three-way interactions respectively.}

\item{n.sims}{Number of desired simulated models.}

\item{method}{Format method to present results. The default is set to predictions of log-odds ="odds". For results on the scale of the predictors ="coeffs".}

\item{matched.data}{Additional data set of pre-processed data to compare the same random models on both data sets.}
}
\value{
Returns a data.frame composed by the list of estimated treatment effects and the model used in each iteration. In the case of including an additional data set, the data.frame will have to columns with the respective estimates.
}
\description{
Takes in a binary dependent data set and outputs a data.frame of the estimated average treatment effects.
}
\examples{
\dontrun{
peacekeeping <- read.table("peacekeeping.csv")
p <- c("untype4", "wartype", "logcost", "wardur", "factnum", "trnsfcap", "decade")

mlmd(peacekeeping, "pbs2s3", "untype4", predictors=p level=2, method="odds")}
}
