% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(Dat, norm = NULL)
}
\arguments{
\item{Dat}{A Dataset object.}

\item{norm}{Either a string indicating the variable name from
Map to use for normalization, or a numeric vector with normalization
values}
}
\description{
Takes a dataset object and normalizes the abundance
matrix by a variable from Map or a vector provided.
}

