% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsmd.R
\name{olsmd}
\alias{olsmd}
\title{Ordinary Least Squares Model Dependence Simulations}
\usage{
olsmd(data, dep, treat, predictors, level, n.sims, matched.data = NULL)
}
\arguments{
\item{data}{Data set to perform the analysis.}

\item{dep}{The dependent outcome binary variable that will be used to measure the effect (e.g. "success").}

\item{treat}{The treatment variable (e.g. "treatment").}

\item{predictors}{Vector of predictor variables.}

\item{level}{Level of allowed interactions. Level = (1, 2, 3) and correspond to one-way, two-way and three-way interactions respectively.}

\item{n.sims}{Number of desired simulated models.}

\item{matched.data}{Additional data set of pre-processed data to compare the same random models on both data sets.}
}
\value{
Returns a data.frame composed by the list of estimated treatment effects and the model used in each iteration. In the case of including an additional data set, the data.frame will have to columns with the respective estimates.
}
\description{
Takes in a data set and outputs a data.frame of the estimated average treatment effects.
}
\examples{
\dontrun{
data(lalonde)
p <- c("treat", "age", "black", "hisp", "re75", "married", "educ")

olsmd(lalonde, "re78", "treat", predictors=p level=2, method="odds")}
}
