% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg.r
\name{plotgg.PCA}
\alias{plotgg.PCA}
\title{Plot a PCA}
\usage{
\method{plotgg}{PCA}(x, components = c("PC1", "PC2"), shape = NULL,
  col = NULL, fill = NULL, biplot = FALSE, biplot_color = "grey21",
  point_size = 2)
}
\arguments{
\item{x}{A \code{PCA} object.}

\item{components}{Vector of length 2 indicating which components to plot.}

\item{shape}{String indicating which variable to use as aestetics
mapping for shape. Must correspond to a column header in the Map attribute
of the PCA object.}

\item{col}{String indicating which variable to use as aestetics mapping
for color. Must correspond to a column header in the Map attribute of the
PCA object.}

\item{fill}{String indicating which variable to use as aestetics mapping for
fill. Must correspond to a column header in the Map attribute of the PCA object.}

\item{biplot}{Logical indicating whether the loadings should be plotted as well.}

\item{biplot_color}{Color to use for the loadings in a biplot}

\item{point_size}{size for the points in the plot}
}
\value{
A ggplot2 object of the PCA plot.
}
\description{
Function for plotting results of \code{\link{PCA}}.
}
\examples{
data(Rhizo)
data(Rhizo.map)
Dat <- create_dataset(Rhizo,Rhizo.map)
Dat.pca <- PCA(Dat)

plotgg(PCA(Dat$Tab),point_size=6)
plotgg(Dat.pca,point_size=4)
plotgg(Dat.pca,shape="fraction",point_size=3)
plotgg(Dat.pca,col="accession")
plotgg(Dat.pca,col="accession",shape="fraction",point_size=4,biplot=TRUE)
p1 <- plotgg(Dat.pca,col="accession",components=c("PC2","PC3"),shape="fraction",biplot=TRUE,biplot_color="pink",point_size=6)
p1
}
\author{
Sur from Dangl Lab.
}
\seealso{
\code{\link{PCA}}
}

