% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg.r
\name{plotgg}
\alias{plotgg}
\title{Make ggplot2 plot}
\usage{
plotgg(...)
}
\description{
Generic plot function for ggplot2
}
\author{
Sur fron Dangl Lab
}
\seealso{
\code{\link{plotgg.PCA}}, \code{\link{plotgg.PCO}},
\code{\link{plotgg.site.diversity}}
}

