% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmd.r
\name{plotmd}
\alias{plotmd}
\title{Data Visualization Tool for Model Dependence Simulations}
\usage{
plotmd(data)
}
\arguments{
\item{data}{The data.frame returned by mlmd() or olsmd().}
}
\value{
Returns a density plot of the estimates calculated by mlmd() or olsmd().
}
\description{
A tool to turn the outputted data.frame from mlmd() and olsmd() into a density plot. The function automatically recognizes whether an additional data set (matched.data) was included when running mlmd() or olsmd() and genereates the plot accordingly.
}
