% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_glm.r
\name{print.summary.bootglm}
\alias{print.summary.bootglm}
\title{Print summary.bootglm object}
\usage{
\method{print}{summary.bootglm}(x)
}
\arguments{
\item{x}{A summary.bootglm object}
}
\description{
Print summary.bootglm object
}
\author{
Sur Herrera Paredes
}
