% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.r
\name{rarefaction}
\alias{rarefaction}
\alias{rarefaction.default}
\alias{rarefaction.Dataset}
\title{Rarefaction}
\usage{
rarefaction(x, sample)

\method{rarefaction}{default}(x, sample)

\method{rarefaction}{Dataset}(x, sample)
}
\arguments{
\item{x}{Either a matrix object or a Dataset object.
If a matrix is passed, samples must be the columns
and rows the species.}

\item{sample}{Sample size, either a single integer
value or a vector of the same length the number of
columns (samples) of x.}
}
\value{
The defualt method returns a matrix with
the rarefied counts. The Dataset method returns a
Dataset object where the Tab element is the rarefied
counts.
}
\description{
Performs rarefaction on count table.
}
\details{
Modified from the rrarefy function of vegan-2.0-4.
}
\examples{
data(Rhizo)
data(Rhizo.map)
Dat <- create_dataset(Rhizo,Rhizo.map)
set.seed(712) # Always specify and save the seed before rarefaction
rarefaction(x=Dat$Tab,sample=100)
set.seed(712)
rarefaction(x=Dat,sample=100)$Tab
}
\author{
Sur Herrera Paredes
}
