% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxons.r
\name{remove_taxons}
\alias{remove_taxons}
\title{Remove taxa from a Dataset object.}
\usage{
remove_taxons(Dat, taxons)
}
\arguments{
\item{Dat}{A Dataset object.}

\item{taxons}{String vector indicating
which taxa to remove. Taxon IDs must be
equal to the row names of the abundance
table on Dat.}
}
\value{
A Dataset object.
}
\description{
Remove taxa from a Dataset object.
}
\examples{
data(Rhizo)
data(Rhizo.map)
Dat <- create_dataset(Rhizo, Rhizo.map)
Dat <- remove_samples(Dat,
                      colnames(Dat$Tab)[1:50])
Dat <- remove_taxons(Dat,
                     row.names(Dat$Tab)[1:115])
sum(colSums(Dat$Tab) == 0)
sum(rowSums(Dat$Tab) == 0)

Dat$Tab
Dat.clean <- clean(Dat)
Dat.clean$Tab
}
\seealso{
\link{remove_samples}
}
\author{
Sur Herrera Paredes
}
