% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.r
\name{samples}
\alias{samples}
\title{Get sample names}
\usage{
samples(Dat)
}
\arguments{
\item{Dat}{A dataset object}
}
\value{
A vector of sample names
}
\description{
Get names of samples in Dataset
}
\seealso{
\code{\link{create_dataset}}, \code{\link{taxa}},
\code{\link{variables}}
}
