% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.Dataset.r
\name{subset.Dataset}
\alias{subset.Dataset}
\title{Subset samples from a Dataset object.}
\usage{
\method{subset}{Dataset}(x, subset, drop = FALSE, clean = FALSE)
}
\arguments{
\item{x}{A \code{Dataset} object.}

\item{subset}{Expression indicating samples to subset. See subset in \link{subset}}

\item{drop}{Logical indicating whether to drop variable levels that disappear.}

\item{clean}{Logical indicating whether to use the \link{clean} function on the resulting
dataset prior to returning it.}
}
\value{
A Dataset object
}
\description{
Takes a Dataset object and a logical expression and returns a Dataset with only the samples
that meet the logical expression criteria
}
\details{
If the logical expression evaluation returns an NA value, it is considered equivalent to
FALSE
}
\author{
Sur from Dangl Lab
}

