% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_glm.r
\name{summary.bootglm}
\alias{summary.bootglm}
\title{Summarizing bootstrap_glm model fits}
\usage{
\method{summary}{bootglm}(object, sortby = "Variable", ...)
}
\arguments{
\item{object}{A bootglm object}

\item{sortby}{How to sort the summarized results}
}
\value{
A summary.bootglm object
}
\description{
Summarizing bootstrap_glm model fits
}
\author{
Sur Herrera Paredes
}
