% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_glm.r
\name{summary.matrix.glm}
\alias{summary.matrix.glm}
\title{Summarize matrix GLM results}
\usage{
\method{summary}{matrix.glm}(object, sortby = "Variable")
}
\arguments{
\item{object}{A matrix.glm or bootglm object.}

\item{sortby}{A string indicating how to sort the coefficients.
Either by "Vatiable" or "Taxon"}
}
\value{
Returns a summary.matrix.glm or summary.bootglm
object which is a list containing the following elements:
\describe{
\item{coefficients}{Coefficient table. For matrix.glm objects it
includes standard errors, z-values and p-values. NOTE: Since only
p-values from z-tests are provided at this point, they are only
appropriate for binomial, poisson and negative.binomial models.
For boot.glm objects in includes the 2.5% and 97.5% percentiles
and the bootstrap confidence for each parameter (Pr(param > 0)).}
\item{call}{Original matrix_glm() or bootstrap_glm() call}}
}
\description{
Takes a matrix.glm or bootglm object and reformats the results
}
\author{
Sur Herrera Paredes
}
