% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOR.r
\name{taxa}
\alias{taxa}
\title{Get taxa names}
\usage{
taxa(Dat)
}
\arguments{
\item{Dat}{A dataset object}
}
\value{
A vector of taxa names
}
\description{
Get names of taxa in Dataset
}
\seealso{
\code{\link{create_dataset}}, \code{\link{variables}},
\code{\link{samples}}
}

