% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_blackbox}
\alias{theme_blackbox}
\title{Blackbox theme}
\usage{
theme_blackbox(border.size = 3)
}
\arguments{
\item{border.size}{Thickness of the margin.}
}
\value{
A \code{theme gg} object.
}
\description{
ggplot2 theme that places a black margin on every
panel
}
\author{
Sur Herrera Paredes
}
