% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOR.r
\name{variables}
\alias{variables}
\title{Get metadata variables}
\usage{
variables(Dat)
}
\arguments{
\item{Dat}{A dataset object}
}
\value{
A vector of variable names
}
\description{
Get names of variables in metadata
}
\seealso{
\code{\link{create_dataset}}, \code{\link{taxa}},
\code{\link{samples}}
}

