% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurable_taxa.r
\name{measurable_taxa}
\alias{findGoodOTUs}
\alias{measurable_taxa}
\alias{measurable_taxa.Dataset}
\alias{measurable_taxa.default}
\title{Identify measurable taxa}
\usage{
measurable_taxa(...)

\method{measurable_taxa}{default}(Tab, min_reads_otu, min_samples_otu,
  method = "absolute", table = TRUE)

\method{measurable_taxa}{Dataset}(Dat, min_reads_otu, min_samples_otu,
  method = "absolute", table = TRUE, clean = TRUE)
}
\arguments{
\item{Tab}{Count table matrix. Samples as columns and OTUs as rows.}

\item{method}{Either "absolute" (min absolute number of reads in min
number of sample) or "AbsProp" (absolute number if reads and proportion
of samples).}

\item{table}{logical indicating whether an abundance table should be
returned}

\item{Dat}{A \code{Dataset} object.}

\item{clean}{logical, indicating whether the function \code{\link{clean}}
should be applied to the result.}
}
\value{
When table = FALSE, it returns a logical index vector that shows which taxa pass the
thresholds.

When table = TRUE, it returns a matrix or Dataset object (depending on the method used),
where samples that do not pass the thresholds have been removed
}
\description{
Identfies taxa above certain abundance and prevalence
thresholds, that can be quantitatively analyzed
}
\author{
Sur from Dangl Lab.
}

