% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurable_taxa.r
\name{measurable_taxa}
\alias{findGoodOTUs}
\alias{measurable_taxa}
\alias{measurable_taxa.Dataset}
\alias{measurable_taxa.default}
\title{Identify measurable taxa}
\usage{
measurable_taxa(...)

\method{measurable_taxa}{default}(Tab, min_reads_otu = 25,
  min_samples_otu = 5, method = "absolute", table = TRUE)

\method{measurable_taxa}{Dataset}(Dat, min_reads_otu = 25,
  min_samples_otu = 5, method = "absolute", table = TRUE, clean = TRUE)
}
\arguments{
\item{Tab}{Count table matrix. Samples as columns and OTUs as rows.}

\item{method}{Either "absolute" (min absolute number of reads in min
number of sample) or "AbsProp" (absolute number if reads and proportion
of samples).}

\item{table}{logical indicating whether an abundance table should be
returned}

\item{Dat}{A \code{Dataset} object.}

\item{clean}{logical, indicating whether the function \code{\link{clean}}
should be applied to the result.}
}
\value{
When table = FALSE, it returns a logical index vector that shows which taxa pass the
thresholds.

When table = TRUE, it returns a matrix or Dataset object (depending on the method used),
where samples that do not pass the thresholds have been removed
}
\description{
Identfies taxa above certain abundance and prevalence
thresholds, that can be quantitatively analyzed
}
\details{
Implments abundance filters defined in Benson et al.(2010) & Lundberg et al. (2012).
It is recommended to threshold any Dataset before any quantitative analysis.
}
\author{
Sur from Dangl Lab.
}
\references{
1. Benson AK, Kelly S a, Legge R, Ma F, Low SJ, Kim J, et al. Individuality in gut microbiota
composition is a complex polygenic trait shaped by multiple environmental and host genetic
factors. Proc Natl Acad Sci U S A. 2010 Nov 2;107(44):18933–8. Available from:
\url{http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2973891&tool=pmcentrez&rendertype=abstract}

2. Lundberg DS, Lebeis SL, Herrera Paredes S, Yourstone S, Gehring J, Malfatti S, et al.
Defining the core Arabidopsis thaliana root microbiome. Nature. Nature Publishing Group;
2012 Aug 1;488(7409):86–90. Available from: \url{http://www.nature.com/doifinder/10.1038/nature11237}
}

