% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.r
\name{PCA}
\alias{PCA}
\alias{PCA.default}
\alias{PCA.Dataset}
\title{Principal Component Analysis}
\usage{
PCA(x, cor, dim)

\method{PCA}{default}(x, cor = FALSE, dim = min(nrow(x), ncol(x)))

\method{PCA}{Dataset}(x, cor = FALSE, dim = min(nrow(x$Tab), ncol(x$Tab)))
}
\arguments{
\item{x}{Numeric matrix where samples are columns and rows are species, or a
Dataset object, see \code{\link{create_dataset}}.}

\item{cor}{logical value indicating whether the correlation matrix should
be used instead of the covariance matrix.}

\item{dim}{Number of dimensions to return.}
}
\value{
A \code{PCA} object. Includes the same attributes as a
\code{\link{pca}} object from the \code{\link{labdsv}} package.
When the Dataset method is used, it includes two additional slots:
\itemize{
\item{"Map"}{The Mapping file for the samples.}
\item{"Tax"}{The Taxonomic information of the taxa.}}
}
\description{
Function that performs principal component analysis on an abundance matrix.
}
\details{
This function is the same as function \code{\link{pca}} from the
\code{\link{labdsv}} package, but includes a methdod for Dataset objects.
}
\examples{
data(Rhizo)
data(Rhizo.map)
Dat <- create_dataset(Rhizo,Rhizo.map)
Dat.pca <- PCA(Dat)
plotgg(Dat.pca,col="accession",shape="fraction",point_size=4,biplot=TRUE)
summary(Dat.pca)
}
\seealso{
\link{create_dataset}, \link{pca}, \link{PCO}, \link{pco},
\link{plotgg.pca}
}
\author{
Sur from Dangl Lab.
}
