% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogram.r
\name{phylogram}
\alias{phylogram}
\alias{phylogram.default}
\alias{phylogram.Dataset}
\title{Phylogram}
\usage{
phylogram(...)

\method{phylogram}{default}(Tab, Map = NULL, facet = NULL,
  colname = "Sample", variable.name = "Taxon", value.name = "Abundance",
  scales = "free_x", space = "free_x", nrow.legend = 20, ntaxa = NULL,
  other_name = "other")

\method{phylogram}{Dataset}(Dat, facet = NULL, colname = "Sample",
  variable.name = "Taxon", value.name = "Abundance", scales = "free_x",
  space = "free_x", nrow.legend = 20, ntaxa = NULL,
  other_name = "other")
}
\arguments{
\item{Tab}{A matrix object representing samples as columns and taxa as rows.}

\item{Map}{A data frame with metadata for Tab. One row per sample, and the
rows must be named to match the column names in Tab, and they must be in the same order.}

\item{facet}{facet formula for facet_grid.}

\item{colname}{Label to be used in the x-axis of the phylogram.}

\item{variable.name}{Label to be used in for the colors in the phylogram}

\item{value.name}{Label for the y-axis of the phylogram.}

\item{scales}{scales option for \link{facet_grid}.}

\item{space}{space option for \link{facet_grid}}

\item{nrow.legend}{number of rows to use in the legend.}

\item{ntaxa}{Number of taxa to plot, the rest will be collapsed into
one category. Samples will be sorted by abundance and the top ntaxa will be plotted.}

\item{other_name}{Name to give to the collapsed taxa when ntaxa is used.}

\item{Dat}{A Dataset object.}
}
\value{
A ggplot2 object of the plot.
}
\description{
Make a relative abundance barplot of taxon abundances.
}
\examples{
data(Rhizo)
data(Rhizo.map)
data(Rhizo.tax)
Dat <- create_dataset(Rhizo,Rhizo.map,Rhizo.tax)

Dat.phyl <- collapse_by_taxonomy(Dat = Dat, level = 4)
phylogram(Tab = Dat.phyl$Tab,
          Map = Dat.phyl$Map,
          facet = ~ fraction)
phylogram(Dat.phyl, facet = ~ fraction)
phylogram(Dat.phyl, facet = ~ fraction,
          ntaxa = 5)
}
\author{
Sur Herrera Paredes
}
