% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg_rankabun.r
\name{plotgg_rankabun}
\alias{plotgg_rankabun}
\alias{plotgg_rankabun.default}
\alias{plotgg_rankabun.Dataset}
\title{Make rank abundance plot}
\usage{
plotgg_rankabun(...)

\method{plotgg_rankabun}{default}(Tab, Map = NULL, groupby = NULL,
  sortby = NULL, theme = theme_blackbox(), variable.name = "Taxon",
  value.name = "Abundance", FUN = mean)

\method{plotgg_rankabun}{Dataset}(Dat, groupby = NULL, sortby = NULL,
  theme = theme_blackbox(), variable.name = "Taxon",
  value.name = "Abundance", FUN = mean)
}
\arguments{
\item{Tab}{A matrix object with samples as columns and taxa as rows.}

\item{Map}{A data frame with metadata for Tab. Each row must be a
sample with row names matching column names in Tab and in the same order as in Tab.}

\item{groupby}{Variable name to be used for grouping samples before plotting th rank abundance.}

\item{sortby}{Variable value to be used as reference for sorting the taxa in the rank abundance plot.}

\item{theme}{ggplot2 theme to be used for plotting.}

\item{variable.name}{x-axis label in the plot.}

\item{value.name}{y-axis label in the plot.}

\item{FUN}{Function to use to aggregate samples according to groupby}

\item{Dat}{A Dataset object.}
}
\value{
A ggplot2 plot.
}
\description{
Make a rank abundance plot from a Dataset object.
}
\examples{
data(Rhizo)
data(Rhizo.map)
data(Rhizo.tax)
Dat <- create_dataset(Rhizo,Rhizo.map,Rhizo.tax)

plotgg_rankabun(Tab = Dat$Tab)
plotgg_rankabun(Tab = Dat$Tab, Map = Dat$Map,groupby = "fraction",sortby = "Soil")

plotgg_rankabun(Dat)
plotgg_rankabun(Dat, groupby = "fraction", sortby = "E")
}
\seealso{
\link{plotgg_rankabun2}
}
\author{
Sur Herrera Paredes
}
