% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg_taxon.r
\name{plotgg_taxon}
\alias{plotgg_taxon}
\alias{plotgg_taxon.default}
\alias{plotgg_taxon.Dataset}
\title{Plot single taxon.}
\usage{
plotgg_taxon(...)

\method{plotgg_taxon}{default}(Tab, Map, taxon, x, col = NULL,
  var.name = "Abundance", theme = theme_blackbox)

\method{plotgg_taxon}{Dataset}(Dat, taxon, x, col = NULL,
  var.name = "Abundance", theme = theme_blackbox)
}
\arguments{
\item{Tab}{Matrix with samples as columns and taxa as rows.}

\item{Map}{Data frame with metadata for Tab. Each row must
correspond to a sample in Tab, they must be in the same order,
and be named with the same names as the columns in Tab}

\item{taxon}{String indicating which taxon to plot.}

\item{x}{String indicating which variable to use as the x-axis
in the plot. Should correspond to a header name in Map}

\item{col}{String indicating which variable to use to color the
plot. Can be the same or different as x. It should correspond to
a header name in Map.}

\item{var.name}{y-axis label on the plot.}

\item{theme}{A ggplot2 theme to be used woth the plot.}

\item{Dat}{A Dataset object.}
}
\value{
A ggplot2 plot
}
\description{
Make a box and whiskers plot of a single taxon, while grouping by a variable.
}
\examples{
data(Rhizo)
data(Rhizo.map)
data(Rhizo.tax)
Dat <- create_dataset(Rhizo,Rhizo.map,Rhizo.tax)

# Passing independently an abundance table and mapping file
plotgg_taxon(Tab = Dat$Tab, Map = Dat$Map,
             taxon = "OTU_14834", x = "fraction")
plotgg_taxon(Tab = Dat$Tab, Map = Dat$Map,
             taxon = "OTU_14834", x = "fraction",
             col = "accession")

# Passing a dataset object
plotgg_taxon(Dat, taxon = "OTU_14834",
             x = "fraction")
plotgg_taxon(Dat, taxon = "OTU_14834",
             x = "fraction",
             col = "accession")
}
\seealso{
\link{plotgg_var}
}
\author{
Sur Herrera Paredes
}
