% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg_var.r
\name{plotgg_var}
\alias{plotgg_var}
\alias{plotgg_var.default}
\alias{plotgg_var.Dataset}
\title{Plot single variable.}
\usage{
plotgg_var(...)

\method{plotgg_var}{default}(Map, var.name, x, col = NULL,
  theme = theme_blackbox())

\method{plotgg_var}{Dataset}(Dat, var.name, x, col = NULL,
  theme = theme_blackbox())
}
\arguments{
\item{Map}{A data.frame where each row must correspond
to a sample, and the row names represent sample IDs.}

\item{var.name}{String indicating which variable to plot.
Should correspond to a header name in Map}

\item{x}{String indicating which variable to use as the
x-axis in the plot. Should correspond to a header name in Map}

\item{col}{String indicating which variable to use to
color the plot. Can be the same or different as x. It should
correspond to a header name in Map.}

\item{theme}{A ggplot2 theme to be used woth the plot.}

\item{Dat}{A Dataset object}
}
\value{
A ggplot2 plot.
}
\description{
Make a box and whiskers plot of a single variable,
while grouping by a variable.
}
\examples{
data(Rhizo)
data(Rhizo.map)
data(Rhizo.tax)
Dat <- create_dataset(Rhizo,Rhizo.map,Rhizo.tax)
Dat$Map$Richness <- colSums(Dat$Tab > 0)

# Passing directly a data_frame
plotgg_var(Dat$Map, var.name = "Richness",
           x = "fraction")
plotgg_var(Dat$Map, var.name = "Richness",
           x = "fraction",col = "accession")

# Passing directly a Dataset object
plotgg_var(Dat, var.name = "Richness",
           x = "fraction")
plotgg_var(Dat, var.name = "Richness",
           x = "fraction",
           col = "accession")
}
\author{
Sur Herrera Paredes
}
