% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_richness.r
\name{total_richness}
\alias{total_richness}
\title{Calculate richness of an abundance table.}
\usage{
total_richness(Tab)
}
\arguments{
\item{Tab}{A numeric matrix with columns as samples and rows as taxa.}
}
\value{
A number representing the total richness.
}
\description{
Given a numeric matri representign an abundance
table, it calculates the total richness (ie
species number) in it.
}
\seealso{
\link{site_diversity}, \link{compare_site_diversity}
}
\author{
Sur Herrera Paredes
}
